backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0
werkzeug==3.0.1
openpyxl==3.1.2
pandas==2.1.4
reportlab==4.0.7
PyPDF2==3.0.1
gunicorn==21.2.0
